backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
python-dotenv==1.0.1
redis==5.0.8
rq==1.15.1
pydantic==2.9.2
